fix: 修复 PostgreSQL created_at 默认值类型转换问题

+++ server.js
@@ -21,7 +21,7 @@ async function initDB() {
         category TEXT NOT NULL,
         date TEXT NOT NULL,
         note TEXT,
-        created_at INTEGER DEFAULT (EXTRACT(EPOCH FROM NOW()))
+        created_at INTEGER DEFAULT (EXTRACT(EPOCH FROM NOW())::INTEGER)
       )
     `);
     console.log('Database initialized');
